requirements.txt
groq
flask
requests
python-dotenv